+++ hqporner_api/__init__.py
@@ -1,5 +1,5 @@
-__all__ = ["Client", "BaseCore", "errors", "locals", "functions", "consts", "Video",
+__all__ = ["Client", "BaseCore", "errors", "locals", "consts", "Video",
            "Category", "Sort", "InvalidCategory", "InvalidURL", "InvalidActress"]
 
-from hqporner_api.modules import locals, errors, consts, functions
+from hqporner_api.modules import locals, errors, consts
 from hqporner_api.api import Client, BaseCore, Video, Category, Sort, InvalidCategory, InvalidURL, InvalidActress